adminpanel/requirements.txt
psycopg2
Flask-SQLAlchemy
Flask-Admin
Flask
Flask-Babel
PyYAML